Switch to size_t instead of u32

+++ include/LTEngine/assets/asset_manager.hpp
@@ -18,8 +18,8 @@ namespace LTEngine {
 
 		const std::vector<u8> load(const std::string &path);
 		virtual const std::vector<u8> loadPure(const std::string &path) = 0;
-		void saveAsset(const std::string &path, const u8 *data, u32 size);
-		virtual void saveAssetPure(const std::string &path, const u8 *data, u32 size) = 0;
+		void saveAsset(const std::string &path, const u8 *data, size_t size);
+		virtual void saveAssetPure(const std::string &path, const u8 *data, size_t size) = 0;
 
 		void clearCache() { m_cache.clear(); }
