+++ include/LTEngine/assets/ltbundle_asset_manager.hpp
@@ -11,7 +11,7 @@ namespace LTEngine {
 		~LTBundleAssetManager() = default;
 
 		const std::vector<u8> loadPure(const std::string &path) override;
-		void saveAssetPure(const std::string &path, const u8 *data, u32 size) override;
+		void saveAssetPure(const std::string &path, const u8 *data, size_t size) override;
 
 	private:
 		std::string m_bundlePath;
